DocumentPanel updates
Moved DocumentPanel class to the panel sub-package. Added flexible
hide/show mechanism for table columns in the document panel. Fixed a
minor bug with the size of the scroll pane when the part of the window
was resized. Fixed a minor bug with the statement table, where the
filter was reset after selecting a statement.