Fixed bugs related to attributes and popup windows
If a variable value was not present as an attribute vector, the popup
combo box representing the value was of type String, rather than
AttributeVector. This caused an error when trying to cast the String as
an AttributeVector while saving the contents of the popup combo box.

Also fixed bug #73 reported by aleipold. With a new dna database, the
empty strings were not present in the attribute list, thus causing a
problem while opening a popup because there were no attributes to choose
from. Now adding empty strings to the attribute lists of all short text
variables when creating a new database.