Fixed an issue with rDNA's excludeValues argument
The rNetwork function in the Exporter class accepts String vectors, but
R only submits a vector if there is more than one value. The solution:
if the user submits only a single value to exclude, rDNA's dna_network
function duplicates the value to coerce it into a vector, and the
Exporter class only processes values that haven't been added before,
effectively removing the duplicate again.