Efficiency gains by avoiding complex statement loading
- When selecting a statement from the text, the Stream API is now used
to pre-filter the statements from the statement table, and only the one
selected statement is fetched from the database to reduce overhead.
- Avoiding statement table refresh when statements are deleted using the
statement remove action or the trash can symbol inside a statement popup
window. Instead, the statement table is manually adjusted without
reloading from the database.
- Avoiding statement table refresh when the coder is switched. Instead,
a change is fired in the statement table.
- Avoiding statement table refresh when new statements are added using
the context menu in the text area or the copy symbol in statement popup
windows. Instead, the existing information in memory is used to inject
the statement into the statement table at the appropriate sorting
position. For this, the statement table model and the Statement class
were optimized.