FileChooser class with MacOS compatibility.
Up to (and including) DNA 3.0.9, DNA relied on the JFileChooser class to select files, for example in the dialog windows to open databases, create new databases, import documents from folders or other databases, open and save connection profiles, save log messages to XML or JSON, and export networks. While this works well on Windows and Linux, the security requirements of MacOS are too restrictive. Apple does not allow files to be displayed by the file chooser dialog (other than in Dropbox folders and the Pictures folder). In this update, I have consolidated all instances where a file chooser is used into a new FileChooser class and added a case selection to determine what operating system is being used. If it is MacOS, the native FileDialog is used instead of JFileChooser. This should allow Mac users to open and save files properly. The update fixes #244 and addresses parts of #201.